pyproject.toml
[project]
name = "startup-builder"
version = "0.1.0"
description = "Multi-Agent Startup Builder"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.1",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.9",
]

[dependency-groups]
dev = [
    "google-generativeai>=0.8.6",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
asyncio_mode = "auto"
